#120 - Search field dropdown menu is broken (again)
The useEffect hook that was opening/closing the dropdown was immediately re-opening the dropdown as soon as it was closed, making it look like it just stayed open. It did this because `open` was also one of its dependencies, even though we were interested in changes to options in order to open the dropdown when new options arrived. Turns out we didn't need this useEffect at all. Instead we just strategically placed openDropdown() inside the appropriate event callbacks and that solved it.

The menu selection was also behind by one item when we were putting the current menu selection into the input on arrow up/down. Did some extensive refactoring to this logic and added a fix that simplifies things a bit.