Refactored scroll wheel zoom event again
Using the mouse scroll wheel on the map causes it to zoom, in which case we want to fetch new listing data. This event has been difficult to deal with because the Google Maps zoom event can also be triggered programmatically, rather than just be user interactions with the map. The last of several attempts to handle this was to set a flag to track when a zoom event should be ignored, i.e., "fitBoundsInProgress". This worked well enough but I don't like the solution because code that works with the flag is scattered around too many places, making it hard to read. Adding a native wheel event to the map div is much simpler to read and understand.