Moved layout code for SearchResults & ListingMap
Moved all the CSS and logic needed for layout of SearchResults & ListingMap components out of main layout and into the respective components. The goal is to push client component code further down into the leaf nodes of the component tree now that we're using the app router. Renamed viewType to mobileViewType to make it clear that it's only for mobile breakpoints.