Better accessibility for Price
Added the role group to the parent element and put the label there instead. This makes the whole component more coherent when looking at it in the accessibility tree by nesting the other roles inside the group and giving it a name. Did something similar for the menu button. Hopefully this will make things easier for people using assistive devices. The other bonus is that it makes it easier to select these components using Testing Library.