Fixed bug with searchQueryOptions
We were previously using toString() with a ReadonlyURLSearchParams, which worked fine, but now that we have switched to an object this no longer serializes it correctly. It seems that we don't really need to serialize the object. Its properties won't be changed because it has been frozen with Object.freeze() and the only time its reference will change is if the underlying searchParams changes. The docs also mention that the order of keys doesn't matter.