Replaced useEffect with useMemo for search state
We didn't really need useEffect here. I tried just putting the code to compute searchParamsState directly in the function and the outcome was the same. it seems like it may be prudent to use useMemo here since there are some things going on that are somewhat computationally expensive, so I added that.