Refactored ListingMarker
Using memo was unnecessary, so I was able to remove a lot of code related to that. Keeping the listing object and event listener setup out of the useEffect hook that creates the marker prevents the marker from flickering by avoiding dependencies that would cause it to be re-created. Added even listener hooks to handle the event listeners. These are based on hook use in the react-google-maps library.